Extract display and wiring modules from CLI main.rs
Split carina-cli/src/main.rs into three modules to improve
maintainability:

- display.rs: Plan display functions (print_plan, format_effect,
  format_list_of_maps, format_list_diff)
- wiring.rs: Provider wiring and orchestration (provider_factories,
  get_schemas, validation wrappers, create_plan_from_parsed, etc.)

Also inlined trivial one-liner delegates (convert_backend_config,
find_state_bucket_resource, resource_to_state) at their call sites.

main.rs reduced from ~3900 to ~3090 lines (-808 lines).
