+++ carina-cli/src/display.rs
@@ -1,0 +1,601 @@
+use std::collections::{HashMap, HashSet};
+
+use colored::Colorize;
+
+use carina_core::deps::get_resource_dependencies;
+use carina_core::effect::Effect;
+use carina_core::plan::Plan;
+use carina_core::resource::Value;
+use carina_core::value::{format_value, format_value_with_key, is_list_of_maps, map_similarity};
+
+pub fn print_plan(plan: &Plan) {
+    if plan.is_empty() {
+        println!("{}", "No changes. Infrastructure is up-to-date.".green());
+        return;
+    }
+
+    // Build dependency graph from effects
+    let mut binding_to_effect: HashMap<String, usize> = HashMap::new();
+    let mut effect_deps: HashMap<usize, HashSet<String>> = HashMap::new();
+    let mut effect_bindings: HashMap<usize, String> = HashMap::new();
+
+    for (idx, effect) in plan.effects().iter().enumerate() {
+        let (resource, deps) = match effect {
+            Effect::Create(r) => (Some(r), get_resource_dependencies(r)),
+            Effect::Update { to, .. } => (Some(to), get_resource_dependencies(to)),
+            Effect::Replace { to, .. } => (Some(to), get_resource_dependencies(to)),
+            Effect::Read { resource } => (Some(resource), get_resource_dependencies(resource)),
+            Effect::Delete { .. } => (None, HashSet::new()),
+        };
+
+        if let Some(r) = resource {
+            let binding = r
+                .attributes
+                .get("_binding")
+                .and_then(|v| match v {
+                    Value::String(s) => Some(s.clone()),
+                    _ => None,
+                })
+                .unwrap_or_else(|| r.id.to_string());
+            binding_to_effect.insert(binding.clone(), idx);
+            effect_bindings.insert(idx, binding);
+        }
+        effect_deps.insert(idx, deps);
+    }
+
+    // Build reverse dependency map (who depends on this resource)
+    let mut dependents: HashMap<usize, Vec<usize>> = HashMap::new();
+    for idx in 0..plan.effects().len() {
+        dependents.insert(idx, Vec::new());
+    }
+
+    for (idx, deps) in &effect_deps {
+        for dep in deps {
+            if let Some(&dep_idx) = binding_to_effect.get(dep) {
+                dependents.get_mut(&dep_idx).unwrap().push(*idx);
+            }
+        }
+    }
+
+    // Find root resources (no dependencies within the plan)
+    let mut roots: Vec<usize> = Vec::new();
+    for (idx, deps) in &effect_deps {
+        let has_dep_in_plan = deps.iter().any(|d| binding_to_effect.contains_key(d));
+        if !has_dep_in_plan {
+            roots.push(*idx);
+        }
+    }
+    roots.sort();
+
+    println!("{}", "Execution Plan:".cyan().bold());
+    println!();
+
+    // Track printed effects to avoid duplicates
+    let mut printed: HashSet<usize> = HashSet::new();
+
+    fn print_effect_tree(
+        idx: usize,
+        plan: &Plan,
+        dependents: &HashMap<usize, Vec<usize>>,
+        printed: &mut HashSet<usize>,
+        indent: usize,
+        is_last: bool,
+        prefix: &str,
+    ) {
+        if printed.contains(&idx) {
+            return;
+        }
+        printed.insert(idx);
+
+        let effect = &plan.effects()[idx];
+        let colored_symbol = match effect {
+            Effect::Create(_) => "+".green().bold(),
+            Effect::Update { .. } => "~".yellow().bold(),
+            Effect::Replace { lifecycle, .. } => {
+                if lifecycle.create_before_destroy {
+                    "+/-".magenta().bold()
+                } else {
+                    "-/+".magenta().bold()
+                }
+            }
+            Effect::Delete { .. } => "-".red().bold(),
+            Effect::Read { .. } => "<=".cyan().bold(),
+        };
+
+        // Build the tree connector (shown before child resources)
+        let connector = if indent == 0 {
+            "".to_string()
+        } else if is_last {
+            format!("{}└─ ", prefix)
+        } else {
+            format!("{}├─ ", prefix)
+        };
+
+        // Base indentation for this resource
+        let base_indent = "  ";
+        // Attribute indentation (4 spaces from resource line)
+        let attr_base = "    ";
+
+        match effect {
+            Effect::Create(r) => {
+                println!(
+                    "{}{}{} {} \"{}\"",
+                    base_indent,
+                    connector,
+                    colored_symbol,
+                    r.id.display_type().cyan().bold(),
+                    r.id.name.white().bold()
+                );
+                // Attribute prefix aligns with the resource content
+                let attr_prefix = if indent == 0 {
+                    format!("{}{}", base_indent, attr_base)
+                } else {
+                    let continuation = if is_last {
+                        format!("{}   ", prefix)
+                    } else {
+                        format!("{}│  ", prefix)
+                    };
+                    format!("{}{}   ", base_indent, continuation)
+                };
+                let mut keys: Vec<_> = r
+                    .attributes
+                    .keys()
+                    .filter(|k| !k.starts_with('_'))
+                    .collect();
+                keys.sort();
+                for key in keys {
+                    let value = &r.attributes[key];
+                    if is_list_of_maps(value) {
+                        println!("{}{}:", attr_prefix, key);
+                        println!("{}", format_list_of_maps(value, &attr_prefix));
+                    } else {
+                        println!(
+                            "{}{}: {}",
+                            attr_prefix,
+                            key,
+                            format_value_with_key(value, Some(key)).green()
+                        );
+                    }
+                }
+            }
+            Effect::Update { id, from, to, .. } => {
+                println!(
+                    "{}{}{} {} \"{}\"",
+                    base_indent,
+                    connector,
+                    colored_symbol,
+                    id.display_type().cyan().bold(),
+                    id.name.yellow().bold()
+                );
+                let attr_prefix = if indent == 0 {
+                    format!("{}{}", base_indent, attr_base)
+                } else {
+                    let continuation = if is_last {
+                        format!("{}   ", prefix)
+                    } else {
+                        format!("{}│  ", prefix)
+                    };
+                    format!("{}{}   ", base_indent, continuation)
+                };
+                let mut keys: Vec<_> = to
+                    .attributes
+                    .keys()
+                    .filter(|k| !k.starts_with('_'))
+                    .collect();
+                keys.sort();
+                for key in keys {
+                    let new_value = &to.attributes[key];
+                    let old_value = from.attributes.get(key);
+                    let is_same = old_value
+                        .map(|ov| ov.semantically_equal(new_value))
+                        .unwrap_or(false);
+                    if !is_same {
+                        if is_list_of_maps(new_value) {
+                            println!("{}{}:", attr_prefix, key);
+                            println!("{}", format_list_diff(old_value, new_value, &attr_prefix));
+                        } else {
+                            let old_str = old_value
+                                .map(|v| format_value_with_key(v, Some(key)))
+                                .unwrap_or_else(|| "(none)".to_string());
+                            println!(
+                                "{}{}: {} → {}",
+                                attr_prefix,
+                                key,
+                                old_str.red(),
+                                format_value_with_key(new_value, Some(key)).green()
+                            );
+                        }
+                    }
+                }
+            }
+            Effect::Replace {
+                id,
+                from,
+                to,
+                changed_create_only,
+                lifecycle,
+            } => {
+                let replace_note = if lifecycle.create_before_destroy {
+                    "(must be replaced, create before destroy)"
+                } else {
+                    "(must be replaced)"
+                };
+                println!(
+                    "{}{}{} {} \"{}\" {}",
+                    base_indent,
+                    connector,
+                    colored_symbol,
+                    id.display_type().cyan().bold(),
+                    id.name.magenta().bold(),
+                    replace_note.magenta()
+                );
+                let attr_prefix = if indent == 0 {
+                    format!("{}{}", base_indent, attr_base)
+                } else {
+                    let continuation = if is_last {
+                        format!("{}   ", prefix)
+                    } else {
+                        format!("{}│  ", prefix)
+                    };
+                    format!("{}{}   ", base_indent, continuation)
+                };
+                let mut keys: Vec<_> = to
+                    .attributes
+                    .keys()
+                    .filter(|k| !k.starts_with('_'))
+                    .collect();
+                keys.sort();
+                for key in keys {
+                    let new_value = &to.attributes[key];
+                    let old_value = from.attributes.get(key);
+                    let forces_replacement = changed_create_only.contains(key);
+                    let is_same = old_value
+                        .map(|ov| ov.semantically_equal(new_value))
+                        .unwrap_or(false);
+                    if !is_same {
+                        if is_list_of_maps(new_value) {
+                            let suffix = if forces_replacement {
+                                format!(" {}", "(forces replacement)".magenta())
+                            } else {
+                                String::new()
+                            };
+                            println!("{}{}:{}", attr_prefix, key, suffix);
+                            println!("{}", format_list_diff(old_value, new_value, &attr_prefix));
+                        } else {
+                            let old_str = old_value
+                                .map(|v| format_value_with_key(v, Some(key)))
+                                .unwrap_or_else(|| "(none)".to_string());
+                            if forces_replacement {
+                                println!(
+                                    "{}{}: {} → {} {}",
+                                    attr_prefix,
+                                    key,
+                                    old_str.red(),
+                                    format_value_with_key(new_value, Some(key)).green(),
+                                    "(forces replacement)".magenta()
+                                );
+                            } else {
+                                println!(
+                                    "{}{}: {} → {}",
+                                    attr_prefix,
+                                    key,
+                                    old_str.red(),
+                                    format_value_with_key(new_value, Some(key)).green()
+                                );
+                            }
+                        }
+                    }
+                }
+            }
+            Effect::Delete { id, .. } => {
+                println!(
+                    "{}{}{} {} \"{}\"",
+                    base_indent,
+                    connector,
+                    colored_symbol,
+                    id.display_type().cyan().bold(),
+                    id.name.red().bold()
+                );
+            }
+            Effect::Read { resource } => {
+                println!(
+                    "{}{}{} {} \"{}\" {}",
+                    base_indent,
+                    connector,
+                    colored_symbol,
+                    resource.id.display_type().cyan().bold(),
+                    resource.id.name.cyan().bold(),
+                    "(data source)".dimmed()
+                );
+            }
+        }
+
+        // Print children (dependents)
+        let children = dependents.get(&idx).cloned().unwrap_or_default();
+        let unprinted_children: Vec<_> = children
+            .iter()
+            .filter(|c| !printed.contains(c))
+            .cloned()
+            .collect();
+
+        // New prefix for children: align with attribute indentation
+        let new_prefix = if indent == 0 {
+            format!("{}  ", attr_base)
+        } else {
+            let continuation = if is_last {
+                format!("{}   ", prefix)
+            } else {
+                format!("{}│  ", prefix)
+            };
+            format!("{}   ", continuation)
+        };
+
+        for (i, child_idx) in unprinted_children.iter().enumerate() {
+            let child_is_last = i == unprinted_children.len() - 1;
+            print_effect_tree(
+                *child_idx,
+                plan,
+                dependents,
+                printed,
+                indent + 1,
+                child_is_last,
+                &new_prefix,
+            );
+        }
+    }
+
+    // Print from roots
+    for (i, root_idx) in roots.iter().enumerate() {
+        print_effect_tree(
+            *root_idx,
+            plan,
+            &dependents,
+            &mut printed,
+            0,
+            i == roots.len() - 1,
+            "",
+        );
+    }
+
+    // Print any remaining effects that weren't reachable from roots
+    // (e.g., circular dependencies or isolated resources)
+    let remaining: Vec<_> = (0..plan.effects().len())
+        .filter(|idx| !printed.contains(idx))
+        .collect();
+    for idx in remaining {
+        print_effect_tree(idx, plan, &dependents, &mut printed, 0, true, "");
+    }
+
+    println!();
+    let summary = plan.summary();
+    let mut parts = Vec::new();
+    if summary.read > 0 {
+        parts.push(format!("{} to read", summary.read.to_string().cyan()));
+    }
+    parts.push(format!("{} to add", summary.create.to_string().green()));
+    parts.push(format!("{} to change", summary.update.to_string().yellow()));
+    if summary.replace > 0 {
+        parts.push(format!(
+            "{} to replace",
+            summary.replace.to_string().magenta()
+        ));
+    }
+    parts.push(format!("{} to destroy", summary.delete.to_string().red()));
+    println!("Plan: {}.", parts.join(", "));
+}
+
+pub fn format_effect(effect: &Effect) -> String {
+    match effect {
+        Effect::Create(r) => format!("Create {}", r.id),
+        Effect::Update { id, .. } => format!("Update {}", id),
+        Effect::Replace { id, lifecycle, .. } => {
+            if lifecycle.create_before_destroy {
+                format!("Replace {} (create-before-destroy)", id)
+            } else {
+                format!("Replace {}", id)
+            }
+        }
+        Effect::Delete { id, .. } => format!("Delete {}", id),
+        Effect::Read { resource } => {
+            format!("Read {}", resource.id)
+        }
+    }
+}
+
+/// Format a list-of-maps for Create effect display (multi-line with + prefix)
+pub fn format_list_of_maps(value: &Value, attr_prefix: &str) -> String {
+    let items = match value {
+        Value::List(items) => items,
+        _ => return format_value(value),
+    };
+    let mut lines = Vec::new();
+    for item in items {
+        if let Value::Map(map) = item {
+            let mut keys: Vec<_> = map.keys().collect();
+            keys.sort();
+            let fields: Vec<String> = keys
+                .iter()
+                .map(|k| format!("{}: {}", k, format_value(&map[*k])))
+                .collect();
+            lines.push(format!(
+                "{}  {} {{{}}}",
+                attr_prefix,
+                "+".green().bold(),
+                fields.join(", ")
+            ));
+        }
+    }
+    lines.join("\n")
+}
+
+/// Format a list-of-maps diff for Update effect display.
+/// Uses content-matched comparison (multiset matching) instead of index-based.
+/// 1. Find exact matches between old and new items
+/// 2. Pair remaining unmatched items by similarity for field-level diffs
+/// 3. Display unchanged, modified (~), added (+), and removed (-) items
+pub fn format_list_diff(old_value: Option<&Value>, new_value: &Value, attr_prefix: &str) -> String {
+    let new_items = match new_value {
+        Value::List(items) => items,
+        _ => return format_value(new_value),
+    };
+    let old_items = match old_value {
+        Some(Value::List(items)) => items,
+        _ => &vec![] as &Vec<Value>,
+    };
+
+    let mut old_matched = vec![false; old_items.len()];
+    let mut new_matched = vec![false; new_items.len()];
+
+    // Phase 1: Find exact matches (semantically equal items)
+    for (ni, new_item) in new_items.iter().enumerate() {
+        for (oi, old_item) in old_items.iter().enumerate() {
+            if !old_matched[oi] && old_item.semantically_equal(new_item) {
+                old_matched[oi] = true;
+                new_matched[ni] = true;
+                break;
+            }
+        }
+    }
+
+    // Collect unmatched items
+    let unmatched_old: Vec<usize> = old_matched
+        .iter()
+        .enumerate()
+        .filter(|(_, m)| !**m)
+        .map(|(i, _)| i)
+        .collect();
+    let unmatched_new: Vec<usize> = new_matched
+        .iter()
+        .enumerate()
+        .filter(|(_, m)| !**m)
+        .map(|(i, _)| i)
+        .collect();
+
+    // Phase 2: Pair unmatched items by similarity (most shared key-value pairs)
+    let mut paired: Vec<(usize, usize)> = Vec::new();
+    let mut paired_old = vec![false; unmatched_old.len()];
+    let mut paired_new = vec![false; unmatched_new.len()];
+
+    for (ui_new, &ni) in unmatched_new.iter().enumerate() {
+        let mut best_oi_idx = None;
+        let mut best_sim = 0usize;
+        for (ui_old, &oi) in unmatched_old.iter().enumerate() {
+            if paired_old[ui_old] {
+                continue;
+            }
+            let sim = map_similarity(&old_items[oi], &new_items[ni]);
+            if sim > best_sim {
+                best_sim = sim;
+                best_oi_idx = Some(ui_old);
+            }
+        }
+        if let Some(ui_old) = best_oi_idx.filter(|_| best_sim > 0) {
+            paired.push((unmatched_old[ui_old], ni));
+            paired_old[ui_old] = true;
+            paired_new[ui_new] = true;
+        }
+    }
+
+    // Remaining truly added/removed items
+    let added: Vec<usize> = unmatched_new
+        .iter()
+        .enumerate()
+        .filter(|(i, _)| !paired_new[*i])
+        .map(|(_, &ni)| ni)
+        .collect();
+    let removed: Vec<usize> = unmatched_old
+        .iter()
+        .enumerate()
+        .filter(|(i, _)| !paired_old[*i])
+        .map(|(_, &oi)| oi)
+        .collect();
+
+    // Phase 3: Build output
+    let mut lines = Vec::new();
+
+    // Show unchanged items (exact matches from new list order)
+    for (ni, new_item) in new_items.iter().enumerate() {
+        if let Value::Map(map) = new_item
+            && new_matched[ni]
+        {
+            let mut keys: Vec<_> = map.keys().collect();
+            keys.sort();
+            let fields: Vec<String> = keys
+                .iter()
+                .map(|k| format!("{}: {}", k, format_value(&map[*k])))
+                .collect();
+            lines.push(format!("{}    {{{}}}", attr_prefix, fields.join(", ")));
+        }
+    }
+
+    // Show modified items (paired by similarity)
+    for &(oi, ni) in &paired {
+        if let (Value::Map(old_map), Value::Map(new_map)) = (&old_items[oi], &new_items[ni]) {
+            let mut keys: Vec<_> = new_map.keys().collect();
+            keys.sort();
+            let fields: Vec<String> = keys
+                .iter()
+                .map(|k| {
+                    let new_v = format_value(&new_map[*k]);
+                    let field_same = old_map
+                        .get(*k)
+                        .map(|ov| ov.semantically_equal(&new_map[*k]))
+                        .unwrap_or(false);
+                    if !field_same {
+                        let old_v = old_map
+                            .get(*k)
+                            .map(format_value)
+                            .unwrap_or_else(|| "(none)".to_string());
+                        format!("{}: {} → {}", k, old_v.red(), new_v.green())
+                    } else {
+                        format!("{}: {}", k, new_v)
+                    }
+                })
+                .collect();
+            lines.push(format!(
+                "{}  {} {{{}}}",
+                attr_prefix,
+                "~".yellow().bold(),
+                fields.join(", ")
+            ));
+        }
+    }
+
+    // Show added items
+    for &ni in &added {
+        if let Value::Map(map) = &new_items[ni] {
+            let mut keys: Vec<_> = map.keys().collect();
+            keys.sort();
+            let fields: Vec<String> = keys
+                .iter()
+                .map(|k| format!("{}: {}", k, format_value(&map[*k])))
+                .collect();
+            lines.push(format!(
+                "{}  {} {{{}}}",
+                attr_prefix,
+                "+".green().bold(),
+                fields.join(", ")
+            ));
+        }
+    }
+
+    // Show removed items
+    for &oi in &removed {
+        if let Value::Map(map) = &old_items[oi] {
+            let mut keys: Vec<_> = map.keys().collect();
+            keys.sort();
+            let fields: Vec<String> = keys
+                .iter()
+                .map(|k| format!("{}: {}", k, format_value(&map[*k])))
+                .collect();
+            lines.push(format!(
+                "{}  {} {{{}}}",
+                attr_prefix,
+                "-".red().bold(),
+                fields.join(", ")
+            ));
+        }
+    }
+
+    lines.join("\n")
+}
